FindInJars - extract `PathMatcher` to not reference it trough `ScalaftmConfig`
